Organise all the files
