+++ public/client-v2/src/components/Feed.js
@@ -6,7 +6,8 @@ import { type Element } from '../services/capiQuery';
 /* eslint-enable import/no-duplicates */
 import ScrollContainer from './ScrollContainer';
 import FeedItem from './FeedItem';
-import FrontsCAPISearchInput from './CAPI/FrontsCAPISearchInput';
+import SearchInput from './FrontsCAPIInterface/SearchInput';
+import Loader from './Loader';
 
 // TODO: get apiKey from context (or speak directly to FrontsAPI)
 
@@ -36,20 +37,45 @@ const getThumbnail = (_elements: Element[]) => {
   return smallestAsset && smallestAsset.file;
 };
 
+type ErrorDisplayProps = {
+  error: ?(Error | string),
+  children: React.Node
+};
+
+const ErrorDisplay = ({ error, children }: ErrorDisplayProps) =>
+  error ? (
+    <div>{typeof error === 'string' ? error : error.message}</div>
+  ) : (
+    children
+  );
+
+type LoaderDisplayProps = {
+  children: React.Node,
+  loading: boolean
+};
+
+const LoaderDisplay = ({ loading, children }: LoaderDisplayProps) =>
+  loading ? <Loader /> : children;
+
 const Feed = () => (
   <ScrollContainer title="Feed">
-    <FrontsCAPISearchInput>
-      {({ response: { results } }) =>
-        results.map(({ webTitle, webUrl, elements }) => (
-          <FeedItem
-            key={webUrl}
-            title={webTitle}
-            href={webUrl}
-            thumbnailUrl={elements && getThumbnail(elements)}
-          />
-        ))
-      }
-    </FrontsCAPISearchInput>
+    <SearchInput>
+      {({ pending, error, value }) => (
+        <ErrorDisplay error={error}>
+          <LoaderDisplay loading={pending}>
+            {value &&
+              value.response.results.map(({ webTitle, webUrl, elements }) => (
+                <FeedItem
+                  key={webUrl}
+                  title={webTitle}
+                  href={webUrl}
+                  thumbnailUrl={elements && getThumbnail(elements)}
+                />
+              ))}
+          </LoaderDisplay>
+        </ErrorDisplay>
+      )}
+    </SearchInput>
   </ScrollContainer>
 );
